Implement quantile regression
This commit adds quantile regression and organizes the code a bit
better. Calculating rolling quantiles is performed at the model
initialization stage, since now they are used multiple times. The
`fit` method is made a bit more abstract to accomodate for the two
fitting methods, MLE or quantile regression. A `main` method
demonstrates a fit for a single case with all currently available
models and methods. This is temporary and will be moved to the right
location in the upcoming commits.