First commit of `fit_models.py`
Right now it's empty. The purpose of this script is to fit all models
to all available cases and store the parameters of the models to files
so that they can be loaded back to produce figures, goodness-of-fit
statistics and other derivative items. `Model` objects will be
modified to support exporting and importing their parameters.