Add code for fitting models to data and for visualization
`fit_models.py`
Fits a model to all archetype-#stories-rc-story-direction cases.
All fitted model objects are gathered and stored using pickle, and
the model parameters are also stored in a Pandas DataFrame (in parquet format).
Currently the bilinear Weibull model is baked into the files, but this
can change in the future, to assess the adequacy of other models too.